+++ src/Presentation/Nop.Web/Nop.Web.csproj
@@ -286,6 +286,8 @@
     <Compile Include="Factories\ProfileModelFactory.cs" />
     <Compile Include="Factories\IReturnRequestModelFactory.cs" />
     <Compile Include="Factories\ReturnRequestModelFactory.cs" />
+    <Compile Include="Factories\IWidgetModelFactory.cs" />
+    <Compile Include="Factories\WidgetModelFactory.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
